Made the package class based and added support for Bruker baf files
Made a base MSIGen class and a subclass for each accepted file extension. What was in the msigen.py file was migrated to base_class.py and msigen.py now contains a function for creating a base class object and dynamically importing the subclass script and initializing the subclass to avoid circular imports. This makes managing parameters easier as they are now class attributes. Additionally, support for MS1 data in Bruker's .baf format has been added. Changes have only been tested on Bruker .baf and .tdf data so far, and the jupyter notebook does not yet reflect these changes.
