+++ MSIGen/__init__.py
@@ -1,7 +1,7 @@
-from MSIGen import msigen
-from MSIGen import raw
+from MSIGen.msigen import msigen
+# from MSIGen import raw
 from MSIGen import tsf
 from MSIGen import visualization
-from MSIGen import D
-from MSIGen import mzml
+# from MSIGen import D
+# from MSIGen import mzml
 from MSIGen import GUI
